File structure for Chart
